+++ smart_dnn/tensor/TensorBase.hpp
@@ -165,10 +165,10 @@ Tensor exp(const Tensor& tensor);
 Tensor log(const Tensor& tensor);
 
 // Reduction operations
-Tensor sum(const Tensor& input, const std::vector<int>& axes = {}, bool keepDims = false);
-Tensor mean(const Tensor& input, const std::vector<int>& axes = {}, bool keepDims = false);
-Tensor max(const Tensor& input, const std::vector<int>& axes = {}, bool keepDims = false);
-Tensor min(const Tensor& input, const std::vector<int>& axes = {}, bool keepDims = false);
+Tensor sum(const Tensor& input, const std::vector<size_t>& axes = {}, bool keepDims = false);
+Tensor mean(const Tensor& input, const std::vector<size_t>& axes = {}, bool keepDims = false);
+Tensor max(const Tensor& input, const std::vector<size_t>& axes = {}, bool keepDims = false);
+Tensor min(const Tensor& input, const std::vector<size_t>& axes = {}, bool keepDims = false);
 Tensor clip(const Tensor& input, const double& min, const double& max);
 
 // Creation functions
